+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - support for 1.8.x dropped
 - double quotes in a host now raises an error
 - calling `require 'addressable'` will now load both the URI and Template files
+- assigning to the `hostname` component with an `IPAddr` object is now supported
 
 # Addressable 2.3.8
 - fix warnings